B2BKing group pricing conditions
Added support for B2BKing group prices. 
Woo product table already worked together with B2BKing in showing the specific prices for the logged in user. However, whenever conditions of min or max were set in Woo product table, the filtering was done on the regular prices (not the B2B-group prices of the current logged-in user). This addition adds support for the group prices whenever B2BKing is active and the logged-in user is in a B2B group. It is as suggested by the B2BKing support team and verified working.
When B2BKing is not installed or the user is not logged-in as a B2B-group user, everything just falls back to the original behaviour.